+++ src/modules/ofdm/ofdm_modulator.h
@@ -17,22 +17,22 @@ private:
     const size_t frame_out_size;
     const size_t data_in_size;
 
-    std::vector<std::complex<float>> prs_fft_ref;
-    std::vector<std::complex<float>> prs_time_ref;
+    std::vector<std::complex<int16_t>> prs_fft_ref;
+    std::vector<std::complex<int16_t>> prs_time_ref;
 
     // frequency domain buffers used for dqpsk encoding
-    std::vector<std::complex<float>> last_sym_fft;
-    std::vector<std::complex<float>> curr_sym_fft;
+    std::vector<std::complex<int16_t>> last_sym_fft;
+    std::vector<std::complex<int16_t>> curr_sym_fft;
 public:
     OFDM_Modulator(
         const OFDM_Params _params, 
-        tcb::span<const std::complex<float>> _prs_fft_ref);
+        tcb::span<const std::complex<int16_t>> _prs_fft_ref);
     ~OFDM_Modulator();
     bool ProcessBlock(
-        tcb::span<std::complex<float>> frame_out_buf, 
+        tcb::span<std::complex<int16_t>> frame_out_buf, 
         tcb::span<const uint8_t> data_in_buf);
 private:
     void CreateDataSymbol(
         tcb::span<const uint8_t> sym_data_in, 
-        tcb::span<std::complex<float>> sym_out);
+        tcb::span<std::complex<int16_t>> sym_out);
 };
